# apps/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Convixx API",
    "description": "Multi-tenant AI Voice Calling Platform backend. RAG-based question answering with self-hosted LLM and OpenAI fallback.\n\n**Documentation:** Interactive UI at `/docs`. Raw OpenAPI JSON at `/docs/json` (same origin as this API).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (recommended for Try it out)"
    },
    {
      "url": "https://convixx.in",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Admin token for customer APIs. Set in .env as ADMIN_TOKEN."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Customer API key. Scopes KB, agents, ask, chat, voice to that customer."
      }
    },
...
  },
  "paths": {
...
    "/voice/cartesia-test-console/config": {
      "get": {
        "tags": [
          "Cartesia"
        ],
        "summary": "Resolve the temp test customer's config (TEMPORARY/DIAGNOSTIC)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "cartesia_configured": true,
                  "customer_id": "...",
                  "agent_id": "...",
                  "api_key": "temp_test_...",
                  "voice_id": "...",
                  "model_id": "sonic-3.5"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Temp test customer not found. Run: npx ts-node scripts/create-temp-voice-test-customer.ts (from apps/api)"
                }
              }
            }
          }
        }
      }
    },
    "/voice/cartesia-test-console/sample-audio/{name}": {
      "get": {
        "tags": [
          "Cartesia"
        ],
        "summary": "Serve a pre-recorded sample WAV for automated testing (TEMPORARY/DIAGNOSTIC)",
        "description": "Debug-only: lets this console be exercised end-to-end without a physical microphone. `name` must match `^[a-zA-Z0-9_-]+\\.wav$`.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]+\\.wav$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WAV audio",
            "content": {
              "audio/wav": {}
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "invalid filename"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "not found"
                }
              }
            }
          }
        }
      }
    },
    "/voice/cartesia-test-console/turn": {
      "post": {
        "tags": [
          "Cartesia"
        ],
        "summary": "Batch turn: STT → runAskPipeline (RAG) → one-shot Cartesia TTS (TEMPORARY/DIAGNOSTIC)",
        "description": "multipart/form-data with a `file` field (recorded WAV). Returns the full answer and audio in one response, with a stage-by-stage timings breakdown. Uses the hardcoded temp test customer regardless of x-api-key value (any key works — the customer is resolved server-side by a marker name).",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "transcript": "Do you offer a free trial?",
                  "answer": "Yes, we offer a 14-day free trial...",
                  "source": "kb-direct",
                  "audio": {
                    "content_type": "audio/mpeg",
                    "bytes": 90324,
                    "base64": "..."
                  },
                  "timings": {
                    "stt_ms": 654,
                    "ask_ms": 161,
                    "tts_ms": 1169,
                    "tts_ttfb_ms": 135,
                    "tts_wall_ms": 1171,
                    "total_ms": 2006
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Multipart field `file` with recorded audio is required."
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Temp test customer not found — run the seed script first."
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "RAG pipeline failed"
                }
              }
            }
          },
          "502": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Cartesia TTS failed"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "CARTESIA_API_KEY is not configured on this server"
                }
              }
            }
          }
        }
      }
    },
    "/voice/cartesia-test-console/turn-stream": {
      "post": {
        "tags": [
          "Cartesia"
        ],
        "summary": "Streaming turn: STT → token-streamed RAG → chunked, concurrent Cartesia TTS via SSE (TEMPORARY/DIAGNOSTIC)",
        "description": "multipart/form-data with a `file` field (recorded WAV). Response is `text/event-stream` (Server-Sent Events), not a single JSON body: events are `transcript`, `retrieval_done`, `first_token`, `text_delta` (repeated), `answer_complete`, `first_audio`, `audio_chunk` (repeated, one per synthesized sentence, in order), `done`, and `error`. Sentence-level TTS chunks synthesize concurrently (capped at 2 in flight, matching this Cartesia account's concurrency limit) while later text is still streaming from the LLM, so audio for the first sentence can start playing before the full answer is generated.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream. Example frames (each `event:`/`data:` pair):\ntranscript {\"transcript\":\"...\",\"stt_ms\":414}\ntext_delta {\"delta\":\"Yes\"}\naudio_chunk {\"index\":0,\"text\":\"...\",\"base64\":\"...\",\"contentType\":\"audio/mpeg\",\"bytes\":95339,\"synthMs\":1127}\ndone {\"total_ms\":6301,\"first_token_ms\":4445,\"first_audio_ms\":6300,\"chunk_count\":1}",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Multipart field `file` with recorded audio is required."
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Temp test customer not found — run the seed script first."
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Temp agent has no cartesia_avatar_id configured."
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "CARTESIA_API_KEY is not configured on this server"
                }
              }
            }
          }
        }
      }
    }
  }
}